+++ modules/stream_manager/builtin/cpp/sscanfBuiltin.cpp
@@ -50,6 +50,7 @@ Nelson::StreamGateway::sscanfBuiltin(int nLhs, const ArrayOfVector& argIn)
                     if (m < 0) {
                         Error(_W("Wrong value >= 0 expected."));
                     }
+                    n = 1;
                 } else {
                     double* ptr = (double*)param3.getDataPointer();
                     m = ptr[0];